+++ source/utils/lcd_controller.h
@@ -10,7 +10,7 @@
 
 void prepareLCD();
 void frequencyLCD(uint8_t freq);
-void reactionLCD(uint8_t time);
+void reactionLCD(uint32_t time);
 void clearLCD();
 
 #endif /* LCD_CONTROLLER_H_ */
